feat(refact): refactored the folder structure
