+++ template/src/navigations/routes.ts
@@ -1,0 +1,2 @@
+export { default as Home } from '@src/screens/main/Home';
+export { default as Setting } from '@src/screens/main/Setting';
