+++ template/src/screens/examples/Query.tsx
